Complete saving and loading. It still doesn't work, but I'll figure it out. Also needs some additional features, but most of it is here.

+++ MacGame/Menus/AlertBoxMenu.cs
@@ -14,7 +14,7 @@ namespace MacGame
 
             ok.Chosen += acceptEvent;
 
-            this.Position = new Vector2(Game1.Camera.ViewPortWidth / 2, Game1.Camera.ViewPortHeight * (1f / 3f));
+            this.Position = new Vector2(Game1.Camera.ViewPortWidth / 2, (int)(Game1.Camera.ViewPortHeight * (1f / 3f)));
 
             this.menuOptions.Add(ok);
             this.IsOverlay = true;
